Identación de código y limpieza.

+++ DatingApp/Controllers/MessagesController.cs
@@ -30,16 +30,10 @@ namespace DatingApp.Controllers
         [HttpGet("{id}", Name = "GetMessage")]
         public async Task<IActionResult> GetMessage(int userId, int id)
         {
-            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
-            {
-                return Unauthorized();
-            }
+            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value)) return Unauthorized();
 
-            Message messageFromRepo = await _repo.GetMessage(id);
-            if (messageFromRepo == null)
-            {
-                return NotFound();
-            }
+            var messageFromRepo = await _repo.GetMessage(id);
+            if (messageFromRepo == null) return NotFound();
 
             return Ok(messageFromRepo);
         }
@@ -47,18 +41,16 @@ namespace DatingApp.Controllers
         [HttpGet]
         public async Task<IActionResult> GetMessagesForUser(int userId, [FromQuery] MessageParams messageParams)
         {
-            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
-            {
-                return Unauthorized();
-            }
+            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value)) return Unauthorized();
 
             messageParams.UserId = userId;
 
-            PagedList<Message> messagesFromRepo = await _repo.GetMessagesForUser(messageParams);
+            var messagesFromRepo = await _repo.GetMessagesForUser(messageParams);
 
-            IEnumerable<MessageToReturnDto> messages = _mapper.Map<IEnumerable<MessageToReturnDto>>(messagesFromRepo);
+            var messages = _mapper.Map<IEnumerable<MessageToReturnDto>>(messagesFromRepo);
 
-            Response.AddPagination(messagesFromRepo.CurrentPage, messagesFromRepo.PageSize, messagesFromRepo.TotalCount, messagesFromRepo.TotalPages);
+            Response.AddPagination(messagesFromRepo.CurrentPage, messagesFromRepo.PageSize, messagesFromRepo.TotalCount,
+                messagesFromRepo.TotalPages);
 
             return Ok(messages);
         }
@@ -66,44 +58,36 @@ namespace DatingApp.Controllers
         [HttpGet("thread/{RecipientId}")]
         public async Task<IActionResult> GetMessageThread(int userId, int recipientId)
         {
-            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
-            {
-                return Unauthorized();
-            }
+            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value)) return Unauthorized();
 
-            IEnumerable<Message> messageFromRepo = await _repo.GetMessagesThread(userId, recipientId);
+            var messageFromRepo = await _repo.GetMessagesThread(userId, recipientId);
 
-            IEnumerable<MessageToReturnDto> messageThread = _mapper.Map<IEnumerable<MessageToReturnDto>>(messageFromRepo);
+            var messageThread = _mapper.Map<IEnumerable<MessageToReturnDto>>(messageFromRepo);
 
             return Ok(messageThread);
         }
 
         [HttpPost]
-        public async Task<IActionResult> CreateMessage(int userId, [FromBody] MessageForCreationDto messageForCreationDto)
+        public async Task<IActionResult> CreateMessage(int userId,
+            [FromBody] MessageForCreationDto messageForCreationDto)
         {
-            User sender = await _repo.GetUser(userId);
+            var sender = await _repo.GetUser(userId);
 
-            if (sender.Id != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
-            {
-                return Unauthorized();
-            }
+            if (sender.Id != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value)) return Unauthorized();
 
             messageForCreationDto.SenderId = userId;
 
-            User recipient = await _repo.GetUser(messageForCreationDto.RecipientId);
+            var recipient = await _repo.GetUser(messageForCreationDto.RecipientId);
 
-            if (recipient == null)
-            {
-                return BadRequest("No se encontró el usuario");
-            }
+            if (recipient == null) return BadRequest("No se encontró el usuario");
 
-            Message message = _mapper.Map<Message>(messageForCreationDto);
+            var message = _mapper.Map<Message>(messageForCreationDto);
 
             _repo.Add(message);
 
             if (await _repo.SaveAll())
             {
-                MessageToReturnDto messageToReturn = _mapper.Map<MessageToReturnDto>(message);
+                var messageToReturn = _mapper.Map<MessageToReturnDto>(message);
                 return CreatedAtRoute("GetMessage", new { userId, id = message.Id }, messageToReturn);
             }
 
@@ -113,32 +97,17 @@ namespace DatingApp.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteMessage(int id, int userId)
         {
-            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
-            {
-                return Unauthorized();
-            }
+            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value)) return Unauthorized();
 
-            Message messageFromRepo = await _repo.GetMessage(id);
+            var messageFromRepo = await _repo.GetMessage(id);
 
-            if (messageFromRepo.SenderId == userId)
-            {
-                messageFromRepo.SenderDeleted = true;
-            }
+            if (messageFromRepo.SenderId == userId) messageFromRepo.SenderDeleted = true;
 
-            if (messageFromRepo.RecipientId == userId)
-            {
-                messageFromRepo.RecipientDeleted = true;
-            }
+            if (messageFromRepo.RecipientId == userId) messageFromRepo.RecipientDeleted = true;
 
-            if (messageFromRepo.SenderDeleted && messageFromRepo.RecipientDeleted)
-            {
-                _repo.Delete(messageFromRepo);
-            }
+            if (messageFromRepo.SenderDeleted && messageFromRepo.RecipientDeleted) _repo.Delete(messageFromRepo);
 
-            if (await _repo.SaveAll())
-            {
-                return NoContent();
-            }
+            if (await _repo.SaveAll()) return NoContent();
 
             throw new Exception("Error al borrar el mensaje");
         }
@@ -146,17 +115,11 @@ namespace DatingApp.Controllers
         [HttpPost("{id}/read")]
         public async Task<IActionResult> MarkMessageAsRead(int userId, int id)
         {
-            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
-            {
-                return Unauthorized();
-            }
+            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value)) return Unauthorized();
 
-            Message message = await _repo.GetMessage(id);
+            var message = await _repo.GetMessage(id);
 
-            if (message.RecipientId != userId)
-            {
-                return Unauthorized();
-            }
+            if (message.RecipientId != userId) return Unauthorized();
 
             message.IsRead = true;
             message.DateRead = DateTime.Now;
