+++ DatingApp/Controllers/PhotosController.cs
@@ -26,27 +26,27 @@ namespace DatingApp.Controllers
         private readonly IOptions<CloudinarySettings> _cloudinaryConfig;
         private readonly Cloudinary _cloudinary;
 
-        public PhotosController(IDatingRepository datingRepository, IMapper mapper, IOptions<CloudinarySettings> cloudinaryConfig)
+        public PhotosController(IDatingRepository datingRepository, IMapper mapper,
+            IOptions<CloudinarySettings> cloudinaryConfig)
         {
             _datingRepository = datingRepository;
             _mapper = mapper;
             _cloudinaryConfig = cloudinaryConfig;
 
-            Account acc = new Account(
+            var acc = new Account(
                 _cloudinaryConfig.Value.CloudName,
                 _cloudinaryConfig.Value.ApiKey,
                 _cloudinaryConfig.Value.ApiSecret);
 
             _cloudinary = new Cloudinary(acc);
-
         }
 
         [HttpGet("{id}", Name = "GetPhoto")]
         public async Task<IActionResult> GetPhoto(int id)
         {
-            Photo photoFromRepo = await _datingRepository.GetPhoto(id);
+            var photoFromRepo = await _datingRepository.GetPhoto(id);
 
-            PhotoForReturnDto photo = _mapper.Map<PhotoForReturnDto>(photoFromRepo);
+            var photo = _mapper.Map<PhotoForReturnDto>(photoFromRepo);
 
             return Ok(photo);
         }
@@ -55,21 +55,18 @@ namespace DatingApp.Controllers
         [HttpPost]
         public async Task<IActionResult> AddPhotoForUser(int userId, [FromForm] PhotoForCreationDto photoForCreationDto)
         {
-            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
-            {
-                return Unauthorized();
-            }
+            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value)) return Unauthorized();
 
-            User userFromRepo = await _datingRepository.GetUser(userId);
+            var userFromRepo = await _datingRepository.GetUser(userId);
 
-            IFormFile file = photoForCreationDto.File;
+            var file = photoForCreationDto.File;
 
-            ImageUploadResult uploadResult = new ImageUploadResult();
+            var uploadResult = new ImageUploadResult();
 
             if (file.Length > 0)
             {
-                Stream stream = file.OpenReadStream();
-                ImageUploadParams uploadParams = new ImageUploadParams()
+                var stream = file.OpenReadStream();
+                var uploadParams = new ImageUploadParams()
                 {
                     File = new FileDescription(file.Name, stream),
                     Transformation = new Transformation().Width(500).Height(500).Crop("fill").Gravity("face")
@@ -81,18 +78,15 @@ namespace DatingApp.Controllers
             photoForCreationDto.Url = uploadResult.Url.ToString();
             photoForCreationDto.PublicId = uploadResult.PublicId;
 
-            Photo photo = _mapper.Map<Photo>(photoForCreationDto);
+            var photo = _mapper.Map<Photo>(photoForCreationDto);
 
-            if (!userFromRepo.Photos.Any(u => u.IsMain))
-            {
-                photo.IsMain = true;
-            }
+            if (!userFromRepo.Photos.Any(u => u.IsMain)) photo.IsMain = true;
 
             userFromRepo.Photos.Add(photo);
 
             if (await _datingRepository.SaveAll())
             {
-                PhotoForReturnDto photoToReturn = _mapper.Map<PhotoForReturnDto>(photo);
+                var photoToReturn = _mapper.Map<PhotoForReturnDto>(photo);
                 return CreatedAtRoute("GetPhoto", new { userId = userId, id = photo.Id }, photoToReturn);
             }
 
@@ -102,34 +96,22 @@ namespace DatingApp.Controllers
         [HttpPost("{id}/setMain")]
         public async Task<IActionResult> SetMainPhoto(int userId, int id)
         {
-            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
-            {
-                return Unauthorized();
-            }
+            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value)) return Unauthorized();
 
-            User user = await _datingRepository.GetUser(userId);
+            var user = await _datingRepository.GetUser(userId);
 
-            if (!user.Photos.Any(p => p.Id == id))
-            {
-                return Unauthorized();
-            }
+            if (!user.Photos.Any(p => p.Id == id)) return Unauthorized();
 
-            Photo photoFromRepo = await _datingRepository.GetPhoto(id);
+            var photoFromRepo = await _datingRepository.GetPhoto(id);
 
-            if (photoFromRepo.IsMain)
-            {
-                return BadRequest("Esta es la foto principal");
-            }
+            if (photoFromRepo.IsMain) return BadRequest("Esta es la foto principal");
 
-            Photo currentMainPhoto = await _datingRepository.GetMainPhotoForUser(userId);
+            var currentMainPhoto = await _datingRepository.GetMainPhotoForUser(userId);
             currentMainPhoto.IsMain = false;
 
             photoFromRepo.IsMain = true;
 
-            if (await _datingRepository.SaveAll())
-            {
-                return NoContent();
-            }
+            if (await _datingRepository.SaveAll()) return NoContent();
 
             return BadRequest("Esta foto no pudo ser la principal");
         }
@@ -137,46 +119,28 @@ namespace DatingApp.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeletePhoto(int userId, int id)
         {
-            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
-            {
-                return Unauthorized();
-            }
+            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value)) return Unauthorized();
 
-            User user = await _datingRepository.GetUser(userId);
+            var user = await _datingRepository.GetUser(userId);
 
-            if (!user.Photos.Any(p => p.Id == id))
-            {
-                return Unauthorized();
-            }
+            if (!user.Photos.Any(p => p.Id == id)) return Unauthorized();
 
-            Photo photoFromRepo = await _datingRepository.GetPhoto(id);
+            var photoFromRepo = await _datingRepository.GetPhoto(id);
 
-            if (photoFromRepo.IsMain)
-            {
-                return BadRequest("Esta foto principal no puede ser eliminada");
-            }
+            if (photoFromRepo.IsMain) return BadRequest("Esta foto principal no puede ser eliminada");
 
             if (photoFromRepo.PublicId != null)
             {
-                DeletionParams deleteParams = new DeletionParams(photoFromRepo.PublicId);
+                var deleteParams = new DeletionParams(photoFromRepo.PublicId);
 
-                DeletionResult result = await _cloudinary.DestroyAsync(deleteParams);
+                var result = await _cloudinary.DestroyAsync(deleteParams);
 
-                if (result.Result == "ok")
-                {
-                    _datingRepository.Delete(photoFromRepo);
-                }
+                if (result.Result == "ok") _datingRepository.Delete(photoFromRepo);
             }
 
-            if (photoFromRepo.PublicId == null)
-            {
-                _datingRepository.Delete(photoFromRepo);
-            }
+            if (photoFromRepo.PublicId == null) _datingRepository.Delete(photoFromRepo);
 
-            if (await _datingRepository.SaveAll())
-            {
-                return Ok();
-            }
+            if (await _datingRepository.SaveAll()) return Ok();
 
             return BadRequest("Ha fallado al eliminar la foto");
         }
